Bug 571147 extend 'lic.net' with common facilities
Abstract 'NetServices' as a general way to define actions supported by a
server
