+++ bundles/org.eclipse.passage.lbc.base/src/org/eclipse/passage/lbc/internal/base/ChoreDraft.java
@@ -20,6 +20,7 @@ import org.eclipse.passage.lbc.internal.base.api.RawRequest;
 import org.eclipse.passage.lic.internal.api.EvaluationInstructions;
 import org.eclipse.passage.lic.internal.api.LicensingException;
 import org.eclipse.passage.lic.internal.net.handle.Chore;
+import org.eclipse.passage.lic.internal.net.handle.Failure;
 import org.eclipse.passage.lic.internal.net.handle.NetResponse;
 
 abstract class ChoreDraft implements Chore {
